Better integration of the login logic into the application instead of creating a separate app for inlog logic all is now in the same app to make sure the code is dry and doest heva 2 different migration files
