--- ramp_up_project/accounts/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-
-from .views import SignUpView
-
-
-urlpatterns = [
-    path("signup/", SignUpView.as_view(), name="signup"),
-]
