+++ ramp_up_project/ramp_up_project/settings.py
@@ -38,7 +38,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'blog.apps.BlogConfig',
-    "accounts",
 ]
 
 MIDDLEWARE = [
