+++ ramp_up_project/ramp_up_project/urls.py
@@ -17,9 +17,12 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 from django.views.generic.base import TemplateView
+
+from blog.views import SignUpView
+
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path("accounts/", include("accounts.urls")),  # new
     path("accounts/", include("django.contrib.auth.urls")),
+    path("signup/", SignUpView.as_view(), name="signup"),
     path("", TemplateView.as_view(template_name="home.html"), name="home"),
 ]
